+++ vms/server/nx_vms_server/src/nx/vms/server/analytics/db/time_period_fetcher.h
@@ -65,7 +65,7 @@ private:
         const Filter& filter,
         const TimePeriodsLookupOptions& options);
 
-    std::set<int> lookupCombinedAttributes(
+    std::set<int64_t> lookupCombinedAttributes(
         nx::sql::QueryContext* queryContext,
         const QString& text);
